fix order update and gdpr2

+++ integrations/esputnik/Esputnik_Contact.php
@@ -87,7 +87,7 @@ class Esputnik_Contact
         $users = $this->get_latest_users_activity();
         if(count($users) > 0){
             foreach ($users as $user_id){
-                $this->delete_from_yespo($user_id);
+                if(!get_user_by( 'ID', $user_id )) $this->delete_from_yespo($user_id);
             }
         }
     }
